+++ source/formats/guitar_pro/document.h
@@ -98,6 +98,21 @@ struct GraceNote
     int myDuration;
 };
 
+struct Bend
+{
+    struct Point
+    {
+        // Percentage of the note's duration where the point is placed.
+        double myOffset = 0;
+        // Percentage of a full step.
+        int myValue = 0;
+    };
+
+    int myBendType = 0;
+    int myBendValue = 0;
+    std::vector<Point> myPoints;
+};
+
 struct Note
 {
     Note(int string);
@@ -115,6 +130,7 @@ struct Note
     std::optional<int> myDynamic;
     std::optional<int> myTrilledFret;
     std::optional<GraceNote> myGraceNote;
+    std::optional<Bend> myBend;
     bool myIsLetRing;
     bool myIsHammerOnOrPullOff;
     bool myHasPalmMute;
@@ -136,7 +152,7 @@ struct Note
 private:
     void loadNoteEffects(InputStream &stream, const Track &track);
     void loadNoteEffectsGp3(InputStream &stream);
-    void loadBend(InputStream &stream);
+    Bend loadBend(InputStream &stream);
     void loadSlide(InputStream &stream);
     void loadHarmonic(InputStream &stream, const Track &track);
 };
@@ -187,6 +203,18 @@ struct Staff
 
 struct Measure
 {
+    struct TimeSignatureChange
+    {
+        int myNumerator = 4;
+        int myDenominator = 4;
+    };
+
+    struct KeySignatureChange
+    {
+        int myAccidentals = 0;
+        bool myIsMinor = false;
+    };
+
     Measure();
     void load(InputStream &stream);
     void loadStaves(InputStream &stream, const std::vector<Track> &tracks);
@@ -194,12 +222,12 @@ struct Measure
     bool myIsDoubleBar;
     bool myIsRepeatBegin;
     /// The numerator and denominator if there is a time signature change,
-    std::optional<std::pair<int, int>> myTimeSignatureChange;
+    std::optional<TimeSignatureChange> myTimeSignatureChange;
     /// If there is a repeat end, this contains the repeat count.
     std::optional<int> myRepeatEnd;
     /// If there is a key change, this contains the key and whether it is
     /// minor.
-    std::optional<std::pair<int8_t, bool>> myKeyChange;
+    std::optional<KeySignatureChange> myKeyChange;
     /// Optional rehearsal sign.
     std::optional<std::string> myMarker;
     /// Optional alternate ending number.